Removed extraneous files, new build.xml not added yet. still on master branch on accident
