--- html/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
- <head>
-   <title>cs56, S11, Mantis Issue 0000341</title>
- </head>
- <body>
-   <h1>cs56, S11, Mantis Issue 0000341</h1>
-   <ul>
-     <li><a href="https://foo.cs.ucsb.edu/56mantis/view.php?id=341">mantis ticket</a></li>
-     <li><a href="download">download</a></li>
-     <li><a href="browse">browse</a></li>
-     <li><a href="javadoc">javadoc</a></li>
-     <li><a href="jws">jws</a></li>
-     <li>Programmer: Dane Pitkin</a></li>
-   </ul>
-<p A GUI implementing a Foreign Vocabulary Quiz for Mantis Choice Points</p>
- </body>
-</html>
