--- jws/.htaccess
@@ -1,1 +1,0 @@
-AddType application/x-java-jnlp-file .jnlp
