--- jws/ForeignVocabQuizGUI.jnlp
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<jnlp spec="0.2 1.0" codebase="http://www.cs.ucsb.edu/~dpitkin/cs56/S11/issues/0000341/jws" href="ForeignVocabQuizGUI.jnlp">
- <information>
-   <title>UCSB CS56 S11 Mantis Ticket 0000341 *Dane Pitkin* ForeignVocabQuizGUI</title>
-   <vendor>UCSB CS56</vendor>
-   <homepage href="." />
-   <description>A Foreign Vocab Quiz GUI</description>
-  <icon href="cs56.png" />
-  <offline-allowed />
- </information>
- <resources>
-  <j2se version="1.3+" />
-  <jar href="cs56_S11_0000341.jar" />
- </resources>
- <application-desc main-class="edu.ucsb.cs56.S11.dpitkin.choice3.ForeignVocabQuizGUI"/>
-</jnlp>
-
-
