--- jws/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
- <title>JWS for Mantis Ticket 0000341, Dane Pitkin for CS56, S11</title>
-</head>
-<body>
- <h1>JWS for Mantis Ticket 0000341</h1>
- <p>*Dane Pitkin* for CS56, S11</p>
- <p>Start the <a href="ForeignVocabQuizGUI.jnlp">ForeignVocabQuizGUI</a>.</p>
-</body>
-</html>
