+++ src/ForeignVocabWord.java
@@ -1,10 +1,8 @@
-package edu.ucsb.cs56.S11.dpitkin.choice3;
-
 /**A vocabulary word that holds both the english and foreign versions of the word.
- *
- * @author Dane Pitkin
- * @version Mantis Ticket 0000273, CS56, S11, choice points 3.
- */
+ *  *
+ *   * @author Dane Pitkin
+ *    * @version Mantis Ticket 0000273, CS56, S11, choice points 3.
+ *     */
 
 public class ForeignVocabWord{
 
@@ -13,9 +11,9 @@ public class ForeignVocabWord{
 
 
     /** Constructor
-     *@param foreign  initializes foreign word
-     *@param english  initializes English word
-     */
+ *      *@param foreign  initializes foreign word
+ *           *@param english  initializes English word
+ *                */
 
     public ForeignVocabWord(String english, String foreign){
 	this.englishWord = english;
@@ -23,7 +21,7 @@ public class ForeignVocabWord{
     }
 
     /**Constructor
-     */
+ *      */
 
     public ForeignVocabWord(){
 	this.englishWord = "";
@@ -31,8 +29,8 @@ public class ForeignVocabWord{
     }
 
     /** Get English word.
-     *@return <code>englishWord</code> of type String.
-     */
+ *      *@return <code>englishWord</code> of type String.
+ *           */
 
     public String getEnglishWord(){
 	return this.englishWord;
@@ -40,33 +38,33 @@ public class ForeignVocabWord{
     
 
     /** Get foreign word.
-     *@return <code>foreignWord</code> of type String.
-     */
+ *      *@return <code>foreignWord</code> of type String.
+ *           */
 
     public String getForeignWord(){
 	return this.foreignWord;
     }
     
     /**Set English word.
-     *@param word of type String.
-     */
+ *      *@param word of type String.
+ *           */
 
     public void setEnglishWord(String word){
 	englishWord = word;
     }
 
     /** Set foreign word.
-     *@param word of type String.
-     */
+ *      *@param word of type String.
+ *           */
     
     public void setForeignWord(String word){
 	foreignWord = word;
     }
 
     /** Compares a word with <code>foreignWord</code> ignoring case.
-     *@param word word to compare.
-     *@return <code>boolean</code> - are the two words equal?
-     */
+ *      *@param word word to compare.
+ *           *@return <code>boolean</code> - are the two words equal?
+ *                */
 
     public boolean equalsForeignWord(String word){
 	return this.foreignWord.equalsIgnoreCase(word);
@@ -74,9 +72,9 @@ public class ForeignVocabWord{
 
 
     /** Compares a word with <code>englishWord</code> ignoring case.
-     *@param word word to compare.
-     *@return <code>boolean</code> - are the two words equal?
-     */
+ *      *@param word word to compare.
+ *           *@return <code>boolean</code> - are the two words equal?
+ *                */
     public boolean equalsEnglishWord(String word){
 	return this.englishWord.equalsIgnoreCase(word);
     }
