--- src/edu/ucsb/cs56/S11/dpitkin/choice3/vocabulary.txt
@@ -1,110 +1,0 @@
-die Gruppe
-musical group
-das Ensemble
-band
-das Publikum
-audience
-das Theater
-theatre
-die Buehne
-stage
-der Musiker
-musician
-der Saenger
-singer
-Der Klavierspieler
-piano player
-der Auftritt
-performance
-auftreten
-to perform
-die Garderobe
-dressing room
-der Anzug
-suit
-die Gage
-salaray for a performer
-der Vertrag
-contract
-der Erfolg
-success
-der Termin
-appointment
-die Aufnahme
-recording
-der Applaus
-applause
-gruenden
-to found
-proben
-to rehearse
-klatschen
-to applaud
-begeistert
-enthusiastic
-juedisch
-jewish
-der Arier
-aryan
-has Hakenkreuz
-swastika
-das Dritte Reich
-Third Reich
-die Vorladung
-citation
-kooperieren
-to cooperate
-der Komponist
-composer
-der Neffe
-nephew
-niederschlagen
-to throw down
-bedrohen
-to threaten
-spucken
-to spit
-verletzt sein
-to be injured
-das Flugblatt
-leaflet
-der Widerstand
-resistance
-das NS-Regime
-National socialist regime
-die Widerstandsbewegung
-resistance movement
-vervielfaeltigen
-to make copies
-die Beschaffung
-obtaining
-die Tarnung
-disguise
-versenden
-to send out
-erschuettern
-to make shaky
-das Bewusstsein
-consciousness
-der Zweifel
-doubt
-die Schuld
-guilt
-vorhalten
-to confront
-der Flur
-corridor
-die Empore
-gallery
-der Lichthof
-light shaft
-der Pedell
-custodian
-verhaften
-to arrest
-festnehmen
-to take into custody
-abwerfen
-to throw down
-die Ausstellung
-exhibition
